+++ persistence/src/test/kotlin/org/agrfesta/sh/api/cache/adapters/AbstractCacheAdapterTest.kt
@@ -17,7 +17,8 @@ import org.springframework.data.redis.core.StringRedisTemplate
     TestContainersConfig::class,
     AbstractCacheAdapterTest.JacksonTestConfig::class,
     RedisCache::class,
-    SensorsCurrentReadingsCacheAdapter::class
+    SensorsCurrentReadingsCacheAdapter::class,
+    DeviceBatteryCacheAdapter::class
 )
 abstract class AbstractCacheAdapterTest {